+++ lib/textus/extensions.rb
@@ -15,8 +15,8 @@ module Textus
       raise UsageError.new("no active registry; extension code must be loaded by a Store")
   end
 
-  def self.fetcher(name, &)
-    current_registry.register_fetcher(name, &)
+  def self.action(name, &)
+    current_registry.register_action(name, &)
   end
 
   def self.reducer(name, &)
@@ -25,5 +25,9 @@ module Textus
 
   def self.hook(event, name, &)
     current_registry.register_hook(event, name, &)
+  end
+
+  def self.doctor_check(name, &)
+    current_registry.register_doctor_check(name, &)
   end
 end